solutions/task3/pullsubscriber/requirements.txt
Flask==3.1.0
google-cloud-pubsub==2.27.3
google-cloud-storage==3.1.0
google-cloud-bigquery==3.30.0
setuptools>=40.8.0
file:vendor/clickstream-0.1.0.dev1.tar.gz